+++ cypress/e2e/fundamentalTypes/all.cy.js
@@ -14,9 +14,6 @@ import './file.cy.js';
 import './filelist.cy.js';
 import './blob.cy.js';
 import './domexception.cy.js';
-import './domrect.cy.js';
-import './dompoint.cy.js';
-import './dommatrix.cy.js';
 import './noneditable.cy.js';
 
 import './BooleanObject.cy.js';