+++ cypress/e2e/superTypes/all.cy.js
@@ -1,3 +1,7 @@
 import './SpecialNumber.cy.js';
 import './SpecialRealNumber.cy.js';
 import './errorsSpecial.cy.js';
+
+import './domrect.cy.js';
+import './dompoint.cy.js';
+import './dommatrix.cy.js';